Fixing issues withviewModels and font references
