+++ Revisadinho/MaintanenceScene/ViewModel/MaintenanceViewModel.swift
@@ -1,0 +1,28 @@
+//
+//  MaintenanceViewModel.swift
+//  Revisadinho
+//
+//  Created by Hiago Chagas on 21/09/21.
+//
+
+import Foundation
+import CoreData
+
+class MaintenanceViewModel {
+    private let model: Model = Model.shared
+
+    func getMaintenancesByDate(startingDate: Date, endingDate: Date, viewContext: NSManagedObjectContext = DatabaseController.shared.viewContext) -> [Maintenance] {
+        var maintenances: [Maintenance] = []
+        let maintenancesDAO = model.getMaintenances(byStartDate: startingDate, toDate: endingDate, viewContext: viewContext)
+        for maintenance in maintenancesDAO {
+            guard let id = maintenance.id else { return []}
+            guard let date = maintenance.date else { return []}
+            let hodometer = maintenance.hodometer
+            let maintenanceItens = maintenance.maintenanceItens?.map { maintenanceItem in
+                return MaintenanceItem(rawValue: Int(maintenanceItem))!
+            }
+            maintenances.append(Maintenance(id: id, hodometer: hodometer, date: date, maintenanceItens: maintenanceItens ?? []))
+        }
+        return maintenances
+    }
+}
